organize model summary utilities
